# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个简单的后端系统 API 文档，包含用户管理、视频管理等功能。",
        "title": "后端系统 API在线测试文档",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Server-lion",
            "url": "https://github.com/Effortful-lion",
            "email": "server-lion@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "39.105.136.3:8888",
    "paths": {
        "/hall": {
            "get": {
                "description": "查询所有演出厅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "演出厅管理"
                ],
                "summary": "查询所有演出厅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "put": {
                "description": "更新演出厅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "演出厅管理"
                ],
                "summary": "更新演出厅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "演出厅信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HallUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "添加演出厅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "演出厅管理"
                ],
                "summary": "添加演出厅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "演出厅信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.HallIsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除演出厅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "演出厅管理"
                ],
                "summary": "删除演出厅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "演出厅id",
                        "name": "hall_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/hall/{hall_id}": {
            "get": {
                "description": "查询所有演出厅",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "演出厅管理"
                ],
                "summary": "查询所有演出厅",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "演出厅id",
                        "name": "hall_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登录接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "全局接口"
                ],
                "summary": "登录接口",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录响应信息",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/play": {
            "get": {
                "description": "获得所有剧目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "剧目管理"
                ],
                "summary": "获得所有剧目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "put": {
                "description": "更新剧目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "剧目管理"
                ],
                "summary": "更新剧目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "剧目更新信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PlayUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "description": "添加剧目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "剧目管理"
                ],
                "summary": "添加剧目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "剧目信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PlayInsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/play/{play_id}": {
            "get": {
                "description": "获得特定剧目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "剧目管理"
                ],
                "summary": "获得特定剧目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "剧目id",
                        "name": "play_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除剧目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "剧目管理"
                ],
                "summary": "删除剧目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "剧目id",
                        "name": "play_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "注册接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "全局接口"
                ],
                "summary": "注册接口",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserSignUpReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册成功",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        },
        "/userinfo": {
            "post": {
                "description": "获取用户信息接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "全局接口"
                ],
                "summary": "获取用户信息接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/resp.ResponseData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.HallIsertReq": {
            "type": "object",
            "required": [
                "hall_col",
                "hall_name",
                "hall_row"
            ],
            "properties": {
                "hall_col": {
                    "description": "列数",
                    "type": "integer"
                },
                "hall_name": {
                    "description": "演出厅名称，唯一",
                    "type": "string"
                },
                "hall_row": {
                    "description": "行数",
                    "type": "integer"
                }
            }
        },
        "dto.HallUpdateReq": {
            "type": "object",
            "required": [
                "hall_col",
                "hall_id",
                "hall_name",
                "hall_row"
            ],
            "properties": {
                "hall_col": {
                    "type": "integer"
                },
                "hall_id": {
                    "type": "integer"
                },
                "hall_name": {
                    "type": "string"
                },
                "hall_row": {
                    "type": "integer"
                }
            }
        },
        "dto.PlayInsertReq": {
            "type": "object",
            "required": [
                "play_description",
                "play_end_time",
                "play_name",
                "play_price",
                "play_start_time"
            ],
            "properties": {
                "play_description": {
                    "type": "string"
                },
                "play_end_time": {
                    "type": "string"
                },
                "play_name": {
                    "type": "string"
                },
                "play_price": {
                    "type": "number"
                },
                "play_start_time": {
                    "type": "string"
                }
            }
        },
        "dto.PlayStatu": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-varnames": [
                "PlayStatusBefore",
                "PlayStatusDuring",
                "PlayStatusAfter"
            ]
        },
        "dto.PlayUpdateReq": {
            "type": "object",
            "required": [
                "play_description",
                "play_end_time",
                "play_id",
                "play_name",
                "play_price",
                "play_start_time",
                "play_statu"
            ],
            "properties": {
                "play_description": {
                    "type": "string"
                },
                "play_end_time": {
                    "type": "string"
                },
                "play_id": {
                    "type": "integer"
                },
                "play_name": {
                    "type": "string"
                },
                "play_price": {
                    "type": "number"
                },
                "play_start_time": {
                    "type": "string"
                },
                "play_statu": {
                    "$ref": "#/definitions/dto.PlayStatu"
                }
            }
        },
        "dto.UserLoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "auth": {
                    "description": "权限，选填",
                    "type": "string"
                },
                "password": {
                    "description": "密码，必填",
                    "type": "string"
                },
                "username": {
                    "description": "用户名，必填",
                    "type": "string"
                }
            }
        },
        "dto.UserSignUpReq": {
            "type": "object",
            "required": [
                "auth",
                "password",
                "re_password",
                "username"
            ],
            "properties": {
                "auth": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "user",
                        "staff",
                        "manager",
                        "ticketor",
                        "finance",
                        "account"
                    ]
                },
                "password": {
                    "description": "密码，必填",
                    "type": "string"
                },
                "re_password": {
                    "description": "确认密码，必填且必须与Password相同",
                    "type": "string"
                },
                "username": {
                    "description": "用户名，必填",
                    "type": "string"
                }
            }
        },
        "resp.ResCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                2000,
                2001,
                2002
            ],
            "x-enum-comments": {
                "CodeError": "服务器内部错误",
                "CodeInvalidParams": "参数错误",
                "CodeInvalidPassword": "密码错误",
                "CodeNeedLogin": "需要登录",
                "CodeNoAuth": "没有权限",
                "CodeSuccess": "成功",
                "CodeUserExist": "用户已存在",
                "CodeUserNotExist": "用户不存在"
            },
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeInvalidParams",
                "CodeError",
                "CodeNeedLogin",
                "CodeNoAuth",
                "CodeUserExist",
                "CodeUserNotExist",
                "CodeInvalidPassword"
            ]
        },
        "resp.ResponseData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "自定义的数据  ,omitempty可忽略空值不展示  Data any \t`json:\"data,omitempty\"`"
                },
                "msg": {
                    "description": "自定义的msg"
                },
                "status": {
                    "description": "自定义的status",
                    "allOf": [
                        {
                            "$ref": "#/definitions/resp.ResCode"
                        }
                    ]
                }
            }
        }
    }
}